+++ segment.h
@@ -4,7 +4,7 @@
 /* base class of arc and spiral                       */
 /*                                                    */
 /******************************************************/
-/* Copyright 2012,2014,2015,2016 Pierre Abbat.
+/* Copyright 2012,2014,2015,2016,2017 Pierre Abbat.
  * This file is part of Bezitopo.
  * 
  * Bezitopo is free software: you can redistribute it and/or modify
@@ -127,7 +127,7 @@ public:
   {
     return 0;
   }
-  virtual int in(xy pnt)
+  virtual double in(xy pnt)
   {
     return 0;
   }
